requirements.txt
Flask==2.1.1
pandas==1.3.3
joblib==1.0.1
imbalanced-learn==0.8.1
google-generativeai==0.2.2
